backend/requirements.txt
# Copyright (C) 2025 HandShake
# Licensed under the Apache License, Version 2.0.
# See LICENSE file for details.

asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
numpy==1.26.2
sqlparse==0.5.1
tzdata==2024.2
djangorestframework-simplejwt==5.2.2
gunicorn==23.0.0


